[main] Java String Tokens
https://www.hackerrank.com

Given a string "S",  matching the regular expression [A-Za-z !,?._'@]+,
split the string into tokens. We define a token to be one or more
consecutive English alphabetic letters.
Then, print the number of tokens, followed by each token on a new line.

1 <= s.length() <= (4 * 10^5)

String s is composed of any of the following:
	English alphabetic letters, blank spaces, exclamation points (!),
	commas (,), question marks (?), periods (.), underscores (_),
	apostrophes ('), and at symbols (@).