[main] Tag Content Extractor
The name of the start and end tags must be same.
The HTML code <h1>Hello World</h2> is not valid, because the text starts
with an h1 tag and ends with a non-matching h2 tag.

Tags can be nested, but content between nested tags is considered not
valid.
For example, in <h1><a>contents</a>invalid</h1>, contents is valid but
invalid is not valid.

Tags can consist of any printable characters.