[main] BigInteger
In this problem, you have to add and multiply huge numbers! These
numbers are so big that you can't contain them in any ordinary data
types like a long integer.

Use the power of Java's BigInteger class and solve this problem.

"a" and "b" are non-negative integers and can have maximum 200 digits.