[main] Java 1D Array (Part 2)
Let's play a game on an array! You're standing at index 0 of an
n-element array named .
From some index i (where 0 <= i < n), you can perform one of the
following moves:
	Move Backward: If cell i-1 exists and contains a 0, you can walk back
to cell i-1.
Move Forward:
	If cell i+1 contains a zero, you can walk to cell i+1.
	If cell i+leap contains a zero, you can jump to cell i+leap.
	If you're standing in cell n-1 or the value of i+leap>=n, you can walk
or jump off the end of the array and win the game.

Complete the canWin function in the editor below.
canWin has the following parameters:
int leap: the size of the leap
int game[n]: the array to traverse

boolean: true if the game can be won, otherwise false